feat: [structured-packages-implementation-rec-04-replay-001] - Own replay behind the Recordings root
Move neutral load/plan/observe replay behavior into the parent-private replay subservice and delegate from the Recordings root composition.
